kijavitottam a hibakat amiket a csabi irt(kommenteztem a reszeket), ha meg lenne barmi ami eszkozlest igenyel legyszi jelezzetek es megcsinalom
