+++ CarRental.Api/Program.cs
@@ -5,10 +5,8 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 
 
-
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddInfrastructure();
@@ -16,21 +14,20 @@ builder.Services.AddInfrastructure();
 builder.Services.AddDbContext<CarRentalDbContext>(options =>
     options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
 
+
+//egyberaktam a 3 Service.AddControllert
+builder.Services.AddControllers()
+    .AddJsonOptions(options => {
+        options.JsonSerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
+        options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
+    });
+
 builder.Services.AddCors(options =>
 {
     options.AddPolicy("AllowFrontend", policy => policy.WithOrigins("http://localhost:5173")
               .AllowAnyMethod()
               .AllowAnyHeader());
 });
-
-builder.Services.AddControllers()
-    .AddJsonOptions(options => {
-        options.JsonSerializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
-    });
-
-builder.Services.AddControllers()
-    .AddJsonOptions(options =>
-        options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter()));
 
 var app = builder.Build();
 
@@ -51,3 +48,8 @@ app.MapControllers();
 app.Run();
 
 //Local: http://localhost:5173/ itt erheto el a cucc
+//ha nem mukodne akkor a developer shellbe kell beirni: cd frontend, npm install majd npm run dev,
+//ha UnathorizedAcces hiba van akkor pedig a win powershellbe kell ez a parancs:
+//Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope CurrentUser
+
+//Set-ExecutionPolicy -ExecutionPolicy Undefined -Scope CurrentUser ezzel pedig vissza lehet allitani a powershellbe ezt az opciot(nem kotelezo)
